refactor: 랭킹표 공통 스타일드 컴포넌트 파일명 수정
