+++ frontend/src/pages/Ranking/RankingTableStyle.ts
